Lay foundation for v0.0.7 of the extension API
